refactor: centralize home page helpers
