--- src/components/Home/helpers/posts.test.ts
@@ -1,389 +1,0 @@
-import { filterFirstSeriesPost } from './posts';
-import { GraphQLResponse } from '../../../types';
-
-const data = {
-  data: {
-    allMarkdownRemark: {
-      edges: [
-        {
-          node: {
-            id: '0363c947-842e-527a-8e48-fe37b559f699',
-            frontmatter: {
-              title: 'E o coronavírus, hein?',
-              subtitle:
-                'Algumas informações básicas e links para se manter informado',
-              series: null,
-              date: '2020-03-13',
-            },
-          },
-        },
-        {
-          node: {
-            id: '2fff163c-8da4-5426-80c2-b6feec58b63a',
-            frontmatter: {
-              title: 'Subindo seu projeto Front-end gratuitamente',
-              subtitle: 'Apresentando: Netlify',
-              series: {
-                id: 'deploy-fe',
-              },
-              date: '2019-03-02',
-            },
-          },
-        },
-        {
-          node: {
-            id: '4b532368-420d-5132-a3a3-5e72ef016f90',
-            frontmatter: {
-              title: 'Coisas que eu não sei em TI e no geral (2018)',
-              subtitle: null,
-              series: null,
-              date: '2019-01-04',
-            },
-          },
-        },
-        {
-          node: {
-            id: '7985b5ff-3736-51da-ab14-0dc6b5bea98e',
-            frontmatter: {
-              title: 'CSS Basics: Position Absolute e Relative',
-              subtitle: null,
-              series: null,
-              date: '2018/01/29',
-            },
-          },
-        },
-        {
-          node: {
-            id: '4e7e3d78-c686-5e44-9253-1f541be4056c',
-            frontmatter: {
-              title: 'Desenvolvedor JavaScript — do Brasil à Europa',
-              subtitle: ' Bye bye Brasil: Planejando o intercâmbio e dicas',
-              series: {
-                id: 'dev-js',
-              },
-              date: '2018-09-12',
-            },
-          },
-        },
-        {
-          node: {
-            id: 'dcec0332-41dd-5642-ad95-a13e4122ade7',
-            frontmatter: {
-              title: 'Inglês na área de TI é obrigatório?',
-              subtitle: 'Reflexões, cenários e dicas de estudo',
-              series: null,
-              date: '2018-07-27',
-            },
-          },
-        },
-        {
-          node: {
-            id: 'a04cf023-e89f-570d-8fc8-e076fd0eba8b',
-            frontmatter: {
-              title: 'Desenvolvedor JavaScript — do Brasil à Europa',
-              subtitle: 'A Jornada do Inglês: Estudos na Irlanda e reflexões',
-              series: {
-                id: 'dev-js',
-              },
-              date: '2018-05-23',
-            },
-          },
-        },
-        {
-          node: {
-            id: 'e289f4b2-5fac-581d-8381-155673b84003',
-            frontmatter: {
-              title: 'Desenvolvedor JavaScript — do Brasil à Europa',
-              subtitle: 'A Jornada do Inglês: Estudando no Brasil',
-              series: {
-                id: 'dev-js',
-              },
-              date: '2018-04-25',
-            },
-          },
-        },
-        {
-          node: {
-            id: '1801d9a9-bbcf-5f15-bf1a-5a66657ca7fe',
-            frontmatter: {
-              title: 'Desenvolvedor JavaScript — do Brasil à Europa',
-              subtitle: 'Entrando pra TI: Networking e Entrevista',
-              series: {
-                id: 'dev-js',
-              },
-              date: '2018-02-16',
-            },
-          },
-        },
-        {
-          node: {
-            id: '0cf33109-28ce-53d0-85af-9ab61f9b0087',
-            frontmatter: {
-              title: 'Desenvolvedor JavaScript — do Brasil à Europa',
-              subtitle: 'Entrando pra TI: O trabalho',
-              series: {
-                id: 'dev-js',
-              },
-              date: '2018-02-16',
-            },
-          },
-        },
-        {
-          node: {
-            id: '283d2169-2fca-5692-80f9-4a6e69fd7528',
-            frontmatter: {
-              title: 'Desenvolvedor JavaScript — do Brasil à Europa',
-              subtitle: 'Estudos específicos, como eu fiz, faço e dicas',
-              series: {
-                id: 'dev-js',
-              },
-              date: '2018-02-16',
-            },
-          },
-        },
-        {
-          node: {
-            id: 'ff42552c-351c-57ca-baf4-d848e2606447',
-            frontmatter: {
-              title: 'Desenvolvedor JavaScript — do Brasil à Europa',
-              subtitle: 'O que você quer ser quando crescer? Pós Ensino Médio',
-              series: {
-                id: 'dev-js',
-              },
-              date: '2018-01-10',
-            },
-          },
-        },
-        {
-          node: {
-            id: '0b5e33d9-f0d9-508d-bfcd-1a9cea79fa35',
-            frontmatter: {
-              title: 'Desenvolvedor JavaScript — do Brasil à Europa',
-              subtitle: 'O que você quer ser quando crescer? Pós Ensino Médio',
-              series: {
-                id: 'dev-js',
-              },
-              date: '2018-01-03',
-            },
-          },
-        },
-        {
-          node: {
-            id: '64012ff5-07b4-5bb2-bf56-b5bff21ce715',
-            frontmatter: {
-              title: 'Desenvolvedor JavaScript — do Brasil à Europa',
-              subtitle: 'O que você quer ser quando crescer? Colegial',
-              series: {
-                id: 'dev-js',
-              },
-              date: '2017-12-22',
-            },
-          },
-        },
-        {
-          node: {
-            id: 'b2871689-9586-5a48-a139-c6aa16009836',
-            frontmatter: {
-              title: 'Desenvolvedor JavaScript — do Brasil à Europa',
-              subtitle: 'Matrix Origins: Onde tudo começou',
-              series: {
-                id: 'dev-js',
-              },
-              date: '2017-12-17',
-            },
-          },
-        },
-        {
-          node: {
-            id: 'ab7faa89-59ad-54c8-978a-0f1be5662e58',
-            frontmatter: {
-              title: 'Desenvolvedor JavaScript — do Brasil à Europa',
-              subtitle: 'Prólogo',
-              series: {
-                id: 'dev-js',
-              },
-              date: '2017-12-11',
-            },
-          },
-        },
-        {
-          node: {
-            id: '0fcc9fcc-9fa4-562a-b544-84e1e84d837f',
-            frontmatter: {
-              title: 'Subindo seu projeto Front-end gratuitamente',
-              subtitle: 'Apresentando: Surge',
-              series: {
-                id: 'deploy-fe',
-              },
-              date: '2017-09-09',
-            },
-          },
-        },
-        {
-          node: {
-            id: '49943c66-240c-5d41-8dcc-f7b9d6afd918',
-            frontmatter: {
-              title: 'Subindo seu projeto Front-end gratuitamente',
-              subtitle: 'Apresentando: Github Pages',
-              series: {
-                id: 'deploy-fe',
-              },
-              date: '2017-08-30',
-            },
-          },
-        },
-        {
-          node: {
-            id: 'e120a497-2e16-5f58-81e0-e390a0cc2aa2',
-            frontmatter: {
-              title: 'Você precisa conhecer o Franz!',
-              subtitle: null,
-              series: null,
-              date: '2017-07-20',
-            },
-          },
-        },
-        {
-          node: {
-            id: '334e11f4-6aef-527c-bc52-ed2a59b38bcf',
-            frontmatter: {
-              title: 'Entendendo de uma vez por todas Expressões Regulares',
-              subtitle: 'Regex no mundo JavaScript',
-              series: {
-                id: 'regex',
-              },
-              date: '2017-07-18',
-            },
-          },
-        },
-        {
-          node: {
-            id: 'ece32b3e-6ac6-5ced-82e6-b85a3b031a39',
-            frontmatter: {
-              title: 'Entendendo de uma vez por todas Expressões Regulares',
-              subtitle: 'Âncoras',
-              series: {
-                id: 'regex',
-              },
-              date: '2017-06-26',
-            },
-          },
-        },
-        {
-          node: {
-            id: 'a7797ce0-0c3d-5d4f-8d23-6a5d1215eab7',
-            frontmatter: {
-              title: 'Entendendo de uma vez por todas Expressões Regulares',
-              subtitle: 'Capturando Grupos',
-              series: {
-                id: 'regex',
-              },
-              date: '2017-06-24',
-            },
-          },
-        },
-        {
-          node: {
-            id: '36485c1e-394c-5b09-a0cd-ae9adfd6a19c',
-            frontmatter: {
-              title: 'Entendendo de uma vez por todas Expressões Regulares',
-              subtitle: 'Quantifiers',
-              series: {
-                id: 'regex',
-              },
-              date: '2017-06-06',
-            },
-          },
-        },
-        {
-          node: {
-            id: 'df5e2155-6076-58b8-b24f-3979cf0a111e',
-            frontmatter: {
-              title: 'Entendendo de uma vez por todas Expressões Regulares',
-              subtitle: 'Classes de Caractere',
-              series: {
-                id: 'regex',
-              },
-              date: '2017-05-29',
-            },
-          },
-        },
-        {
-          node: {
-            id: 'cbb49c91-170f-5d70-835d-28e64f0c0a39',
-            frontmatter: {
-              title: 'Entendendo de uma vez por todas Expressões Regulares',
-              subtitle: 'Metacharacters',
-              series: {
-                id: 'regex',
-              },
-              date: '2017-05-24',
-            },
-          },
-        },
-        {
-          node: {
-            id: 'ad183c81-3acd-560a-99ac-229342a57864',
-            frontmatter: {
-              title: 'Entendendo de uma vez por todas Expressões Regulares',
-              subtitle: 'Introdução',
-              series: {
-                id: 'regex',
-              },
-              date: '2017-05-21',
-            },
-          },
-        },
-      ],
-    },
-  },
-};
-
-describe('fn: filterFirstSeriesPost', () => {
-  it('returns the expected data', () => {
-    expect(filterFirstSeriesPost(data as GraphQLResponse))
-      .toMatchInlineSnapshot(`
-      Array [
-        Object {
-          "node": Object {
-            "frontmatter": Object {
-              "date": "2017-12-11",
-              "series": Object {
-                "id": "dev-js",
-              },
-              "subtitle": "Prólogo",
-              "title": "Desenvolvedor JavaScript — do Brasil à Europa",
-            },
-            "id": "ab7faa89-59ad-54c8-978a-0f1be5662e58",
-          },
-        },
-        Object {
-          "node": Object {
-            "frontmatter": Object {
-              "date": "2017-08-30",
-              "series": Object {
-                "id": "deploy-fe",
-              },
-              "subtitle": "Apresentando: Github Pages",
-              "title": "Subindo seu projeto Front-end gratuitamente",
-            },
-            "id": "49943c66-240c-5d41-8dcc-f7b9d6afd918",
-          },
-        },
-        Object {
-          "node": Object {
-            "frontmatter": Object {
-              "date": "2017-05-21",
-              "series": Object {
-                "id": "regex",
-              },
-              "subtitle": "Introdução",
-              "title": "Entendendo de uma vez por todas Expressões Regulares",
-            },
-            "id": "ad183c81-3acd-560a-99ac-229342a57864",
-          },
-        },
-      ]
-    `);
-  });
-});
