+++ src/pages/index.tsx
@@ -6,7 +6,7 @@ import { graphql } from 'gatsby';
 import Layout from '../components/Layout';
 import { GraphQLResponse } from '../types';
 import AuthorPresentation from '../components/AuthorPresentation';
-import { getAndSanitizePostsFromQueryResponse } from '../helpers/posts';
+import { getAndSanitizePostsFromQueryResponse } from '../components/Home/helpers/posts';
 import { useIntl } from '../context/react-intl';
 import SEO from '../components/SEO';
 import { Filter } from '../components/Home/Filter';
